requirements.txt
black==24.8.0
blinker==1.8.2
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
elastic-transport==8.15.0
elasticsearch==8.15.1
filelock==3.16.1
flake8==7.1.1
identify==2.6.1
idna==3.10
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pre-commit==3.8.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
tomli==2.0.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.6
Werkzeug==3.0.4
